refactor: Rename Spot to Destination
