+++ dristi/lib/src/features/destination/data/repositories/destination_repositories_impl.dart
@@ -1,0 +1,30 @@
+import 'package:dristi/src/features/destination/data/data_sources/destination_data_source.dart';
+import 'package:dristi/src/features/destination/domain/entities/destination_entities.dart';
+import 'package:dristi/src/features/destination/domain/repositories/destination_repositories.dart';
+
+class DestinationRepositoryImp implements DestinationRepository {
+  const DestinationRepositoryImp({
+    required this.dataSource,
+  });
+
+  final DestinationDataSource dataSource;
+
+  @override
+  Future<DestinationEntity> getDestinationData(
+      String appLanguage, String spotId) async {
+    final response = await dataSource.getDestinationData(appLanguage, spotId);
+
+    return response.destination;
+  }
+
+  @override
+  Future<(String, List<DestinationItemsEntity>?)> getDestinationItems(
+      String appLanguage) async {
+    final response = await dataSource.getDestinationItems(appLanguage);
+
+    return Future.value((
+      response.statusMessage!,
+      response.data as List<DestinationItemsEntity>?
+    ));
+  }
+}
