+++ dristi/lib/src/features/destination/presentation/riverpod/destination_data/destination_state.dart
@@ -1,0 +1,30 @@
+import 'package:equatable/equatable.dart';
+
+enum DestinationStatus { initial, loading, success, failure }
+
+class DestinationState<T> extends Equatable {
+  const DestinationState({
+    this.status = DestinationStatus.initial,
+    this.data,
+    this.errorMessage,
+  });
+
+  final DestinationStatus status;
+  final T? data;
+  final String? errorMessage;
+
+  DestinationState copyWith({
+    DestinationStatus? status,
+    T? data,
+    String? errorMessage,
+  }) {
+    return DestinationState(
+      status: status ?? this.status,
+      data: data ?? this.data,
+      errorMessage: errorMessage ?? this.errorMessage,
+    );
+  }
+
+  @override
+  List<Object?> get props => [status, data, errorMessage];
+}
