--- dristi/lib/src/features/spot/domain/repositories/spot_repositories.dart
@@ -1,17 +1,0 @@
-import 'package:dristi/src/features/spot/data/data_sources/spot_data_source.dart';
-import 'package:dristi/src/features/spot/data/repositories/spot_repositories_impl.dart';
-import 'package:dristi/src/features/spot/domain/entities/spot_entities.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-final spotRepositoryProvider = Provider<SpotRepository>(
-  (ref) {
-    final dataSource = ref.read(spotDataSourceProvider);
-    return SpotRepositoryImp(dataSource: dataSource);
-  },
-);
-
-abstract class SpotRepository {
-  Future<SpotEntity> getSpotData(String appLanguage, String spotId);
-
-  Future<(String, List<SpotItemsEntity>?)> getSpotItems(String appLanguage);
-}
