--- dristi/lib/src/features/spot/domain/use_cases/spot_use_cases.dart
@@ -1,26 +1,0 @@
-import 'package:dristi/src/features/spot/domain/entities/spot_entities.dart';
-import 'package:dristi/src/features/spot/domain/repositories/spot_repositories.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-final spotUseCaseProvider = Provider(
-  (ref) {
-    return SpotUseCase(
-      ref.read(spotRepositoryProvider),
-    );
-  },
-);
-
-class SpotUseCase {
-  SpotUseCase(this.repository);
-
-  final SpotRepository repository;
-
-  Future<SpotEntity> getSpotData(String appLanguage, String spotId) async {
-    return repository.getSpotData(appLanguage, spotId);
-  }
-
-  Future<(String, List<SpotItemsEntity>?)> getSpotItems(
-      String appLanguage) async {
-    return repository.getSpotItems(appLanguage);
-  }
-}
