--- dristi/lib/src/features/spot/presentation/riverpod/spot_data/spot_state.dart
@@ -1,30 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-enum SpotStatus { initial, loading, success, failure }
-
-class SpotState<T> extends Equatable {
-  const SpotState({
-    this.status = SpotStatus.initial,
-    this.data,
-    this.errorMessage,
-  });
-
-  final SpotStatus status;
-  final T? data;
-  final String? errorMessage;
-
-  SpotState copyWith({
-    SpotStatus? status,
-    T? data,
-    String? errorMessage,
-  }) {
-    return SpotState(
-      status: status ?? this.status,
-      data: data ?? this.data,
-      errorMessage: errorMessage ?? this.errorMessage,
-    );
-  }
-
-  @override
-  List<Object?> get props => [status, data, errorMessage];
-}
